+++ BlazorRummiSolve.sln
@@ -4,6 +4,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlazorRummiSolve", "BlazorR
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RummiSolve", "RummiSolve\RummiSolve\RummiSolve.csproj", "{57FD7BCB-DE52-4947-8F44-F3E5DCA61D3D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlazorRummiSolve.Tests", "BlazorRummiSolve.Tests\BlazorRummiSolve.Tests.csproj", "{1895341F-D9C3-45B8-AA61-444E26C6BF42}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -18,5 +20,9 @@ Global
 		{57FD7BCB-DE52-4947-8F44-F3E5DCA61D3D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{57FD7BCB-DE52-4947-8F44-F3E5DCA61D3D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{57FD7BCB-DE52-4947-8F44-F3E5DCA61D3D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1895341F-D9C3-45B8-AA61-444E26C6BF42}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1895341F-D9C3-45B8-AA61-444E26C6BF42}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1895341F-D9C3-45B8-AA61-444E26C6BF42}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1895341F-D9C3-45B8-AA61-444E26C6BF42}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
