+++ RummiSolve/RummiSolve/Set.cs
@@ -1,5 +1,3 @@
-using Microsoft.Diagnostics.Tracing.Parsers.Clr;
-
 namespace RummiSolve;
 
 public class Set : ISet
